// src/types/Review.ts
import firebase from "firebase"

export type UserRef = {
    id: string;
    name: string;
}

export type ShopRef = {
    id: string;
    name: string;
}

export type Review = {
    id?: string;
    text: string;
    score: number;
    user: UserRef;
    shop: ShopRef;
    updatedAt : firebase.firestore.Timestamp;
    createdAt : firebase.firestore.Timestamp;
}